Update middleware API in tests
This still works all the way back to 1.11, and brings us closer to the
upstream tests for Django. For now I avoid the response.headers API,
which was only introduced a bit later. The tests for django-brotli are
not yet updated, so I use lambdas as the least invasive diff.